+++ src/main/resources/db/migration/V4__createUser.sql
@@ -1,5 +1,5 @@
 create table user ( id integer, login varchar(15), password VARCHAR(15),
-type smallint);
+type integer);
 
 create sequence user_id_seq start with 1 increment by 1 no minvalue no maxvalue cache 1;
